+++ paparazzi/paparazzi-annotation-processor/src/main/java/app/cash/paparazzi/annotation/processor/models/EnvironmentModel.kt
@@ -1,0 +1,12 @@
+package app.cash.paparazzi.annotation.processor.models
+
+data class EnvironmentModel(
+  val platformDir: String,
+  val appTestDir: String,
+  val resDir: String,
+  val assetsDir: String,
+  val packageName: String,
+  val compileSdkVersion: Int,
+  val platformDataDir: String,
+  val resourcePackageNames: List<String>
+)
